New intermediary GameItem class
