+++ src/engine/CoreModule/Game.ts
@@ -46,39 +46,4 @@ export default class Game {
     // Initialising the player's stats
     this.playerStats = new UserData(1, 3, 2);
   }
-
-  /**
-   * Writes text to the canvas
-   *
-   * @param text - Text to write
-   * @param fontSize - Font size in pixels
-   * @param xCoordinate - Horizontal coordinate in pixels
-   * @param yCoordinate - Vertical coordinate in pixels
-   * @param alignment - Where to align the text
-   * @param color - The color of the text
-   */
-  public writeTextToCanvas(
-    text: string,
-    fontSize: number = 14,
-    xCoordinate: number,
-    yCoordinate: number,
-    alignment: CanvasTextAlign = 'center',
-    color: string = 'white',
-  ): void {
-    this.ctx.font = `${fontSize}px sans-serif`;
-    this.ctx.fillStyle = color;
-    this.ctx.textAlign = alignment;
-    this.ctx.fillText(text, xCoordinate, yCoordinate);
-  }
-
-  /**
-   * Returns a random number between min and max
-   *
-   * @param min - lower boundary
-   * @param max - upper boundary
-   * @returns a random number between min and max
-   */
-  public static randomNumber(min: number, max: number): number {
-    return Math.round(Math.random() * (max - min) + min);
-  }
 }
