+++ src/game/gameItems/Player.ts
@@ -1,10 +1,10 @@
-import GameObject from '../../engine/GameObjectModule/GameObject.js';
+import GameItem from '../../engine/CoreModule/GameItem.js';
 import Transform from '../../engine/ComponentsModule/Transform.js';
 import Mesh from '../../engine/ComponentsModule/Mesh.js';
 import Input from '../../engine/InputModule/Input.js';
 import Vector2 from '../../engine/MathModule/Vector2.js';
 
-export default class Player extends GameObject {
+export default class Player extends GameItem {
   // The speed of the Player movement measured in pixels per second
   public movementSpeed: number;
